yongshan.net v1.0-Beta-7-20250806
yongshan.net v1.0-Beta-7-20250806
We have added our GitHub open source repository, download interface, and Bilibili URL at the top of the website. Now everyone can quickly enter the page to find the information they need.
At the bottom of the website, we have provided the contact information for website developers, which we believe can help you provide more timely feedback to the development team.
In addition, in the music project, we have removed the website's listening feature, which means that if you want to enjoy this song, you must download it locally.